Refactor ComparisonContext to utilize useComparisonProviderState hook, simplifying state management and removing unused functions. This change enhances code clarity and maintainability.
